requirements.txt
requests
folium
python-dotenv
colorama